notifications/dja_admin/requirements.txt
celery==5.4.0
django==5.0.6
gunicorn==22.0.0
python-dotenv==1.0.1
django-split-settings==1.3.1
django-debug-toolbar==4.3.0
django-postgres-extra==2.0.9rc10
django-celery-beat==2.6.0
django-celery-results==2.5.1
psycopg2-binary==2.9.9
jinja2==3.1.4
requests==2.31.0
flower==2.0.1